backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
openai-whisper
openai
python-dotenv
pydantic==2.10.3
aiofiles==24.1.0
gtts==2.5.4
